requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
APScheduler==3.10.4
attrs==23.2.0
certifi==2024.7.4
dnspython==2.6.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pyaes==1.6.1
pymongo==4.8.0
Pyrogram @ https://github.com/KurimuzonAkuma/pyrogram/archive/dev.zip
PySocks==1.7.1
TgCrypto==1.2.5
